# docs/sns/reference/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/sns/posts/with-images": {
      "post": {
        "operationId": "PostImagesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "submissionId",
                  "content",
                  "images"
                ],
                "properties": {
                  "submissionId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "content": {
                    "type": "string",
                    "maxLength": 20000
                  },
                  "visibilityScopeCode": {
                    "type": "string",
                    "enum": [
                      "public",
                      "organization",
                      "followers",
                      "self"
                    ]
                  },
                  "images": {
                    "type": "array",
                    "minItems": 1,
                    "maxItems": 4,
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiSuccess"
                    },
                    {
                      "type": "object",
                      "required": [
                        "success",
                        "data"
                      ],
                      "properties": {
                        "success": {
                          "type": "boolean",
                          "enum": [
                            true
                          ],
                          "example": true
                        },
                        "data": {
                          "$ref": "#/components/schemas/PostDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "본문·사진 개수/크기/형식 오류"
          },
          "401": {
            "description": "로그인 필요"
          },
          "403": {
            "description": "기능 권한 없음"
          },
          "409": {
            "description": "같은 요청의 내용 불일치 또는 삭제된 요청"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "사진과 본문 함께 게시 (동일 요청 재시도 중복 방지)",
        "tags": [
          "sns-post-images"
        ]
      }
    },
    "/api/sns/posts/{postId}/images/{imageId}": {
      "get": {
        "operationId": "PostImagesController_read",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "imageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "이미지 파일",
            "content": {
              "image/jpeg": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "image/png": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "image/webp": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "401": {
            "description": "로그인 필요"
          },
          "403": {
            "description": "기능 권한 없음"
          },
          "404": {
            "description": "이미지 없음 또는 게시물 열람 불가"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "게시물 열람 권한으로 첨부 이미지 조회",
        "tags": [
          "sns-post-images"
        ]
      }
    },
    "/api/sns/posts": {
      "get": {
        "operationId": "PostController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "페이지 번호",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "페이지 크기",
            "schema": {
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "boardId",
            "required": false,
            "in": "query",
            "description": "게시판 ID 필터",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorUserId",
            "required": false,
            "in": "query",
            "description": "작성자 사용자 ID 필터",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "검색어 (제목, 내용)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ],
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/PostDto"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/ApiPaginationMeta"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "서버 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "게시물 목록",
        "tags": [
          "sns-posts"
        ]
      },
      "post": {
        "operationId": "PostController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePostDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiSuccess"
                    },
                    {
                      "type": "object",
                      "required": [
                        "success",
                        "data"
                      ],
                      "properties": {
                        "success": {
                          "type": "boolean",
                          "enum": [
                            true
                          ],
                          "example": true
                        },
                        "data": {
                          "$ref": "#/components/schemas/PostDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "서버 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "게시물 생성",
        "tags": [
          "sns-posts"
        ]
      }
    },
...
  },
  "info": {
    "title": "SSOO SNS API",
    "description": "SNS module API reference",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "ApiErrorDetail": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "INTERNAL_ERROR"
          },
          "message": {
            "type": "string",
            "example": "Internal server error"
          },
          "path": {
            "type": "string",
            "example": "/api/example"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "ApiError": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              false
            ],
            "example": false
          },
          "error": {
            "$ref": "#/components/schemas/ApiErrorDetail"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "success",
          "error"
        ]
      },
      "ApiSuccess": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              true
            ],
            "example": true
          },
          "data": {
            "type": "object",
            "additionalProperties": true
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "success",
          "data"
        ]
      },
      "PostImageDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "fileName": {
            "type": "string"
          },
          "mimeType": {
            "type": "string"
          },
          "fileSize": {
            "type": "string",
            "description": "바이트 수 (문자열)"
          }
        },
        "required": [
          "id",
          "fileName",
          "mimeType",
          "fileSize"
        ]
      },
      "PostDto": {
        "type": "object",
        "properties": {
          "attachments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PostImageDto"
            }
          },
          "id": {
            "type": "string",
            "description": "게시물 ID"
          },
          "authorUserId": {
            "type": "string",
            "description": "작성자 사용자 ID"
          },
          "boardId": {
            "type": "string",
            "description": "게시판 ID"
          },
          "categoryId": {
            "type": "string",
            "description": "카테고리 ID"
          },
          "title": {
            "type": "string",
            "description": "게시물 제목"
          },
          "content": {
            "type": "string",
            "description": "게시물 내용"
          },
          "contentType": {
            "type": "string",
            "description": "콘텐츠 유형"
          },
          "visibilityScopeCode": {
            "type": "string",
            "description": "게시물 공개 범위",
            "enum": [
              "public",
              "organization",
              "followers",
              "self"
            ]
          },
          "targetOrgId": {
            "type": "object",
            "description": "대상 조직 ID"
          },
          "viewCount": {
            "type": "number",
            "description": "조회 수"
          },
          "createdAt": {
            "type": "string",
            "description": "생성일시"
          },
          "updatedAt": {
            "type": "string",
            "description": "수정일시"
          }
        },
        "required": [
          "id",
          "authorUserId",
          "content",
          "contentType",
          "visibilityScopeCode",
          "viewCount",
          "createdAt",
          "updatedAt"
        ]
      },
      "ApiPaginationMeta": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "minimum": 1,
            "example": 1
          },
          "limit": {
            "type": "number",
            "minimum": 1,
            "example": 20
          },
          "total": {
            "type": "number",
            "minimum": 0,
            "example": 100
          }
        },
        "required": [
          "page",
          "limit",
          "total"
        ]
      },
      "CreatePostDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "게시물 제목"
          },
          "content": {
            "type": "string",
            "description": "게시물 내용"
          },
          "contentType": {
            "type": "string",
            "description": "콘텐츠 유형 (text, markdown, html)",
            "default": "text"
          },
          "boardId": {
            "type": "string",
            "description": "게시판 ID"
          },
          "categoryId": {
            "type": "string",
            "description": "카테고리 ID"
          },
          "visibilityScopeCode": {
            "type": "string",
            "description": "게시물 공개 범위",
            "enum": [
              "public",
              "organization",
              "followers",
              "self"
            ],
            "default": "public"
          },
          "tagNames": {
            "description": "태그 이름 목록",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "content"
        ]
      },
...
    }
  }
}